+++ app/pages/logo.py
@@ -4,6 +4,8 @@ from framework.an.fade_out import FadeOut
 
 from config import font_theme
 
+from framework.globe import scrmgr
+
 
 class Logo(Page):
     def __init__(self, rtu: RuntimeUnit):
@@ -13,12 +15,12 @@ class Logo(Page):
             # construct "presented by" label
             StaticLabel(
                 content="presented by", font=font_theme.logo, size=10,
-                centerx=self.x_mean, bottom=self.y_mean,
+                centerx=scrmgr.x_mean, bottom=scrmgr.y_mean,
             ),
             # construct "WERTech" label
             StaticLabel(
                 content="WERTech", font=font_theme.logo, size=20,
-                centerx=self.x_mean, top=self.y_mean,
+                centerx=scrmgr.x_mean, top=scrmgr.y_mean,
             ),
         )
 
@@ -26,4 +28,4 @@ class Logo(Page):
         self.draw_and_flip()
 
         # full-screen fade-out
-        FadeOut(self.rtu, self.screen, click_optional=True, count_down=1000).play()
+        FadeOut(self.rtu, scrmgr.screen, click_optional=True, count_down=1000).play()